# Improve path_parsing performance.